% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{clean_labels_reorder_within}
\alias{clean_labels_reorder_within}
\title{Reorder Within Clean Labeller}
\usage{
clean_labels_reorder_within(labels, sep = "-_-")
}
\arguments{
\item{labels}{A vector of labels given by drlib::reorder_within()
or something similar.}

\item{sep}{The seperator to split the labels on. Can only appear once in
each label. Default is "-_-".}
}
\description{
This is intended to be used with drlib::reorder_within().
It removes the text that drlib::reorder_within() so
that the axis is correctly ordered but the labels are
clean.
}
\seealso{
drlib::reorder_within()
}
