% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{drop_na_verbose}
\alias{drop_na_verbose}
\title{Drop NA verbose}
\usage{
drop_na_verbose(.data, ...)
}
\arguments{
\item{.data}{A data frame or similar object.}

\item{...}{A selection of columns. If empty, all variables are selected. You
can supply bare variable names, select all variables between x and z with
x:z, exclude y with -y. For more options, see the dplyr::select()
documentation.}
}
\description{
A wraper of \code{\link{tidyr::drop_na()}} to drop NA values and give a
warning for how many rows were dropped.
}
\examples{
\dontrun{
set.seed(0)
data <- tibble::tibble(.rows = 40)
data$x1 <- sample(c(NA_real_, 1:4), 40, replace= TRUE)
data$x2 <- sample(c(NA_real_, 1:4), 40, replace= TRUE)
drop_na_verbose(data) # filter any NA column
drop_na_verbose(data, x1) # filter when x1 is NA
drop_na_verbose(data, -x2) # filter when any column, not including x2, is NA
}

}
