% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{filter_verbose}
\alias{filter_verbose}
\title{Filter verbose}
\usage{
filter_verbose(.data, ...)
}
\arguments{
\item{.data}{A data frame or similar object.}

\item{...}{Logical predicates defined in terms of the variables in .data.
Multiple conditions are combined with &. Only rows where the condition
evaluates to TRUE are kept.}
}
\description{
A wraper of \code{\link{dplyr::filter()}} to give a warning for how many rows
were dropped.
}
\examples{
\dontrun{
set.seed(0)
data <- tibble::tibble(.rows = 40)
data$x1 <- sample(c(NA_real_, 1:4), 40, replace= TRUE)
filter_verbose(data, x1 > 0.5)
}

}
