% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{\%||\%}
\alias{\%||\%}
\title{Iff not null}
\usage{
arg1 \%||\% arg2
}
\arguments{
\item{arg1}{the first argument. Will be returned iff
not null.}

\item{arg2}{the second argument.}
}
\description{
Returns the first argument if and only if not null,
else returns the second arguement. Comes from Hadley's
Advanced R: https://adv-r.hadley.nz.
}
\examples{
\dontrun {
x <- NULL
x \%||\% 'default'

x <- 'option1'
x \%||\% 'default'
}

}
