% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{insert_docstring}
\alias{insert_docstring}
\title{Insert Skeleton roxygen2 documentation}
\usage{
insert_docstring()
}
\description{
Inserts skeleton Roxygen2 documentation for a function at the current line if
in a document or in the console if used in the console.
}
\seealso{
\code{\link[docthis]{doc_this}}
}
