% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{new_rmarkdown}
\alias{new_rmarkdown}
\title{New RMarkdown}
\usage{
new_rmarkdown(title, author)
}
\arguments{
\item{title}{optional. The title of the new document.}

\item{author}{optional. The author of the new document.}
}
\description{
Make a new Rmarkdown document with default settings. Users will be prompted
to provide a title and an author if they are not given in the arguments.
}
