% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{quick_color}
\alias{quick_color}
\title{Quick Colors}
\usage{
quick_color(color)
}
\arguments{
\item{color}{A character with the color(s) to return a hex code for.}
}
\description{
This is a way for me to store color hex codes I frequently use for easy recall.
}
