% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{theme_mattle24}
\alias{theme_mattle24}
\title{Mattle24 Theme}
\usage{
theme_mattle24()
}
\description{
This is a custom ggplot2 theme based on \code{\link{ggplot2::theme_bw()}}
developed by @mattle24.
}
