% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{warn_row_difference}
\alias{warn_row_difference}
\title{Warn row difference}
\usage{
warn_row_difference(.data)
}
\arguments{
\item{.data}{A data frame or similar object.}
}
\description{
Helper function to warn the differences in rows. Requires a variable
`starting_rows` to be in the scope.
}
