% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{insert_docstring}
\alias{insert_docstring}
\title{Insert Skeleton Docstring}
\usage{
insert_docstring()
}
\description{
Inserts a skeleton docstring for a function at the current line if in a
document or in the console if used in the console.
}
