% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{theme_mattle24}
\alias{theme_mattle24}
\title{Mattle24 Theme}
\usage{
theme_mattle24(...)
}
\arguments{
\item{...}{Passes on additional arguments to theme()}
}
\description{
This is a custom ggplot2 theme developed by @mattle24.
}
